factored out TokioTimeSeries as tokio.timeseries.TimeSeries
